requirements.txt
asgiref==3.5.2
certifi==2022.5.18.1
charset-normalizer==2.0.12
# !
Django==4.0.4
# !
django-filter==21.1
idna==3.3
# !
requests==2.27.1
sqlparse==0.4.2
urllib3==1.26.9